Retrieve user information for everyone in an edgelist

add_users_data() updates an edgelist created with create_edgelist() by appending user data retrieved with rtweet::lookup_users(). The resulting dataframe keeps all the columns from rtweet but adds "_sender" or "_receiver" to the column names

add_users_data(edgelist)

Arguments

edgelist
An edgelist of senders and receivers, such as that returned by the function create_edgelist().

Value

A dataframe in the form of an edgelist (i.e., with senders and receivers) as well as numerous, appropriately named columns of details about the senders and receivers.

See also

Examples

if (FALSE) {

example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tmp_df <- pull_tweet_data(read_tags(example_url), n = 10)
add_users_data(create_edgelist(tmp_df))
}