Create an edgelist where interaction is defined by quote-tweeting

Starting with a dataframe returned by pull_tweet_data(), get_quotes() pulls out senders and receivers of quote tweets and adds a new column edge_type. get_quotes() is a useful function in itself, but is also used in create_edgelist().

get_quotes(df)

Arguments

df
A dataframe returned by pull_tweet_data()

Value

A dataframe edgelist with column names 'sender', 'receiver', and edge_type, which in this case the edge type is "quote-tweet"

See also

Examples

if (FALSE) {

example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tmp_df <- pull_tweet_data(read_tags(example_url), n = 10)
get_quotes(tmp_df)
}