Retrieve the fullest extent of tweet metadata for more than 90,000 tweets

This function calls pull_tweet_data(), but has a built in delay of 15 minutes to all the Twitter API to reset after looking up 90,000 tweets.

lookup_many_tweets(x, alarm = FALSE)

Arguments

x
A list or vector of tweet ID numbers

alarm
An audible notification that a batch of 90,000 tweets has been completed

Value

A dataframe of tweets and full metadata from the Twitter API