Retrieve geographic coordinates

geocode_tags() retrieves geographic coordinates (i.e., latitude and longitude) based on the locations listed in Twitter user profiles. geocode_tags() pulls from the OpenCage Geocoding API, which requires a OpenCage Geocoding API Key. You can easily secure a key through OpenCage; read more here. We recommend saving your OpenCage Geocoding API Key in the .Renviron file as OPENCAGE_KEY. You can quickly access this file using the R code usethis::edit_r_environ(scope='user'). Add a line to this file that reads: OPENCAGE_KEY="PasteYourOpenCageKeyInsideTheseQuotes". To read your key into R, use the code Sys.getenv('OPENCAGE_KEY'). Note that the geocode_tags() function retrieves your saved API key automatically and securely. Once you've saved the .Renviron file, quit your R session and restart. The function geocode_tags() will work for you from now on.

geocode_tags(df)

Arguments

df

A dataframe or tibble

Value

A tibble of geographic coordinates (i.e., latitude and longitude) that can then be used to plot locations on a map

Details

This function requires authentication; please see vignette("setup", package = "tidytags")

See also

Examples

if (FALSE) {

example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tmp_df <- pull_tweet_data(read_tags(example_url), n = 10)
tmp_geo_coords <- geocode_tags(tmp_df)
tmp_geo_coords

if (requireNamespace("sf", quietly = TRUE)) {
  locations <- sf::st_as_sf(
    tmp_geo_coords,
    coords = c(x = "longitude", y = "latitude"),
    crs = 4326
  )
}

if (requireNamespace("mapview", quietly = TRUE)) {
  mapview::mapview(locations)
}

}