Retrieve geographic coordinates

geocode_tags() retrieves geographic coordinates (i.e., latitude and longitude) based on the locations listed in Twitter user profiles. geocode_tags() pulls from the Google Geocoding API, which requires a Google Geocoding API Key. You can easily secure a key through Google Cloud Platform; read more here. We recommend saving your Google Geocoding API Key in the .Renviron file as **Google_API_key**. You can quickly access this file using the R code usethis::edit_r_environ(scope='user'). Add a line to this file that reads: Google_API_key="PasteYourGoogleKeyInsideTheseQuotes". To read your key into R, use the code Sys.getenv('Google_API_key'). Note that the geocode_tags() function retrieves your saved API key automatically and securely. Once you've saved the .Renviron file, quit your R session and restart. The function geocode_tags() will work for you from now on.

geocode_tags(df, google_key = Sys.getenv("Google_API_key"))

Arguments

df
A dataframe or tibble

google_key
A Google Developers API key which will need to be obtained by the user

Value

Geographic coordinates (i.e., latitude and longitude) which may then be used to plot locations on a map

See also

Examples

  if (FALSE) {
  example_url <- "https://docs.google.com/spreadsheets/d/18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8/edit#gid=8743918"
  tmp_df <- pull_tweet_data(read_tags(example_url), n = 10)
  tmp_processed <- process_tweets(tmp_df)
  tmp_geo_coords <- geocode_tags(tmp_processed)
  tmp_geo_coords
  mapview::mapview(tmp_geo_coords$pnt)
  }