Retrieve the fullest extent of tweet metadata available from the Twitter API

With a TAGS archive imported into R, pull_tweet_data() uses the **rtweet** package to query the Twitter API. Using rtweet requires Twitter API keys associated with an approved developer account. Fortunately, the rtweet vignette, Authentication, provides a thorough guide to obtaining Twitter API keys. We recommend the second suggested method listed in the rtweet vignette, "2. Access token/secret method." Following these directions, you will run the rtweet::create_token() function, which saves your Twitter API keys to the .Renviron file. You can also edit this file directly using the usethis::edit_r_environ(scope='user') function.

pull_tweet_data(df, url_vector = NULL, id_vector = NULL, n = NULL)

Arguments

df
A dataframe of containing the column name 'status_url' (i.e., the hyperlink to specific tweets), such as that returned by read_tags()

url_vector
A vector of tweet URLs, such as as those contained in the 'status_url' column of a dataframe returned by tidytags::read_tags()

id_vector
A vector of tweet ID numbers, such as as those contained in the 'id_str' column of a dataframe returned by tidytags::read_tags()

n
The number of tweets to look up, by default the total number of tweet ID numbers available, but capped at 90,000 due to Twitter API limitations.

Value

A dataframe of tweets and full metadata from the Twitter API

Details

This function requires authentication; please see vignette("setup", package = "tidytags")

This function requires authentication; please see vignette("setup", package = "tidytags")

See also

Examples

if (FALSE) {

example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
tags_content <- read_tags(example_url)
pull_tweet_data(tags_content[1:10, ])
pull_tweet_data(tags_content, n = 10)
pull_tweet_data(url_vector = tags_content$status_url[1:10])
pull_tweet_data(url_vector = tags_content$status_url, n = 10)
pull_tweet_data(id_vector = tags_content$id_str[1:10])
pull_tweet_data(id_vector = tags_content$id_str, n = 10)
pull_tweet_data(url_vector =
  "https://twitter.com/tweet__example/status/1176592704647716864")
pull_tweet_data(id_vector = "1176592704647716864")
}