Retrieve a TAGS archive of tweets and bring into R

Keep in mind that read_tags() uses the **googlesheets4** package, and one requirement is that your TAGS tracker has been "published to the web." To do this, with the TAGS page open in a web browser, navigate to File >> Publish to the web. The Link field should be 'Entire document' and the Embed field should be 'Web page.' If everything looks right, then click the Publish button. Next, click the Share button in the top right corner of the Google Sheets browser window, select Get shareable link, and set the permissions to 'Anyone with the link can view.' The URL needed for R is simply the URL at the top of the web browser, just copy and paste at this point. Be sure to put quotations marks around the URL when entering it into read_tags().

read_tags(url)

Arguments

url
A valid URL (i.e., hyperlink) to a TAGS tracker, or a Google Sheet identifier (i.e., the alphanumeric string following "https://docs.google.com/spreadsheets/d/" in the TAGS tracker's URL.)

Value

A dataframe of the TAGS archive of tweets

See also

Examples

if (FALSE) {

example_url <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
read_tags(example_url)
}