Retrieve a TAGS archive of Twitter statuses and bring into R

Keep in mind that read_tags() uses the googlesheets4 package, and one requirement is that your TAGS tracker has been "published to the web." To do this, with the TAGS page open in a web browser, navigate to File >> Publish to the web. The Link field should be 'Entire document' and the Embed field should be 'Web page.' If everything looks right, then click the Publish button. Next, click the Share button in the top right corner of the Google Sheets browser window, select Get shareable link, and set the permissions to 'Anyone with the link can view.'

read_tags(tags_id, google_key = Sys.getenv("GOOGLE_API_KEY"))

Arguments

tags_id
A Google Sheet identifier (i.e., the alphanumeric string following "https://docs.google.com/spreadsheets/d/" in the TAGS tracker's URL.)

google_key
A Google API key for accessing Google Sheets.

Value

A tibble of the TAGS archive of Twitter statuses

Details

This function requires authentication; please see vignette("setup", package = "tidytags")

See also

Examples

if (FALSE) {

example_tags <- "18clYlQeJOc6W5QRuSlJ6_v3snqKJImFhU42bRkM_OX8"
read_tags(example_tags)
}